+++ examples/center.rs
@@ -11,7 +11,7 @@ use embedded_graphics_simulator::{
 };
 use embedded_text::{
     alignment::CenterAligned,
-    style::{height_mode::HeightMode, TextBoxStyleBuilder},
+    style::{HeightMode, TextBoxStyleBuilder},
     TextBox,
 };