+++ examples/left.rs
@@ -10,7 +10,7 @@ use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, Window,
 };
 use embedded_text::{
-    style::{height_mode::HeightMode, TextBoxStyleBuilder},
+    style::{HeightMode, TextBoxStyleBuilder},
     TextBox,
 };