+++ src/alignment/scrolling.rs
@@ -50,9 +50,7 @@ mod test {
 
     use crate::{
         alignment::Scrolling,
-        style::{
-            height_mode::HeightMode, vertical_overdraw::VerticalOverdraw, TextBoxStyleBuilder,
-        },
+        style::{HeightMode, TextBoxStyleBuilder, VerticalOverdraw},
         utils::test::size_for,
         TextBox,
     };